fix: check `destinations` & `destinationLists` exist
In some configurations, a destinationList or destinations are not
required for the operation of the system. The
`useGetAllDeviceStateFromRoomConfiguration` hook now checks that the
value exists prior to attempting to get the device keys from the lists.
A check was also added for `sourceList`, as there may also be valid
configurations where a `sourceList` is not required.