fix: /system/roomCombinationChanged handling override
For some UIs, it is not necessary to reload the application when
the room combination state changes. In addition, the application may also
be able to update itself without restarting as well, by requesting the
correct room key from Essentials. The handler for the `/system/roomCombinationChanged` now
checks for any registered handlers. If no handlers are found, then the application
is reloaded, which is the existing behavior. If handlers are found, then the handlers
are executed, and the default behavior is not. This allows for customizing what happens
when the room combination state changes, while keeping the existing behavior in place for
backwards compatibility.